+++ lib/src/models/game_record/tournament_game_record.dart
@@ -1,9 +1,15 @@
 import 'package:bridge_game/src/models/game_record/game_record.dart';
 import 'package:bridge_game/src/enums/index.dart';
 import 'package:bridge_game/src/models/card.dart';
+import 'package:json_annotation/json_annotation.dart';
 
+part 'tournament_game_record.g.dart';
+
+@JsonSerializable()
 class TournamentGameRecord extends GameRecord {
   final Card? lead;
+  double scoreIMP = 0;
+  double scoreMP = 0;
   final bool isEmpty;
 
   TournamentGameRecord({
@@ -68,4 +74,11 @@ class TournamentGameRecord extends GameRecord {
       };
     }
   }
+
+  // From JSON
+  factory TournamentGameRecord.fromJson(Map<String, dynamic> json) =>
+      _$TournamentGameRecordFromJson(json);
+
+  // To JSON
+  Map<String, dynamic> toJson() => _$TournamentGameRecordToJson(this);
 }